Now the slow down of the end of walking happens acording to the time the player has walked. Which means, the more time the player has walking the smoother the slow down will be.

+++ source/player/player.h
@@ -67,7 +67,7 @@ class player
         std::array<SDL_Texture*, 3> textures;
 
         // the slow down speed of stopping to walk
-        float slowDownEndWalk;
+        double slowDownEndWalk;
         
         // walking speed
         float walkingSpeed;
